3/home/ubuntu/tradelayer_intelligence_backend/requirements.txt
asgiref==3.9.1
blinker==1.9.0
click==8.2.1
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.2
SQLAlchemy==2.0.41
sqlparse==0.5.3
typing_extensions==4.14.0
uritemplate==4.2.0
Werkzeug==3.1.3
